add delete modal to simplify components and delete children props in modal to add or edit task
