+++ package.json
@@ -15,6 +15,8 @@
     "@types/react-dom": "18.2.6",
     "@types/uuid": "^9.0.2",
     "autoprefixer": "10.4.14",
+    "mobx": "^6.9.0",
+    "mobx-react-lite": "^3.4.3",
     "next": "13.4.9",
     "postcss": "8.4.25",
     "react": "18.2.0",
